Motel Network Design project
Scope of Work: This project will include the design and installation of an AWS network. This project also includes the installation and setup of all the network devices, interfaces, and cables. The following fall under the network devices mentioned: workstations, multi-function devices, servers, switches, rotors, racks, and all related software.

Feasibility Study

The network proposed is designed to serve the Motel, which will consist of 25-30 rooms in 7000 square feet. The building is currently under heavy renovation and practically hallowed. This allows us to have a greater range of flexibility for the installation of the proposed network. The rooms within the building will be a motel, 20 motel rooms, a storage room, a server room, a gym, a lobby, an administrative room, and a break room.

Objectives

Secure Service: The primary goal of this network is to offer secure administrative computing services to Motel. The design ensures that only Motel employees can access it to reduce the possibility of unauthorized use.

Integration and Update: This is a recently built structure with outer walls being the only feature. The new network and equipment being set up will consist of cutting-edge technology tailored for Motel.

Scalability: The scalability of the network design allows for additional network cable equipment to be incorporated as required and when funding permits, without the need to reconfigure the existing network.

Versatile Information Processing: The network will allow users to access, handle, and save information about the motel’s functioning, its customers, finances, invoicing, monitoring, and entrance and exit of the guests staying at the motel.

Infrastructure Required to plan the network coverage based on the size and create heat maps. 7000 total sqft Rooms - 6000 sq ft
Each room is 300 sq ft each
2 floors
20 rooms - 10 rooms on each floor Other than Rooms - 1000 sqft
Lobby
1 Reception counter

1 administration room
Gym
Server room
Break room
Tools used: Figma for group collaboration |